Changed: Merged downcast-helpers.js to downcast-converters.js and upcast-helpers.js to upcast-converters.js.
